refactor(architecture): merge more steps into `core-workflow-steps`
this renames `org.eclipse.sw360.antenna.workflow.processors.enricher.LicenseResolver` to `org.eclipse.sw360.antenna.workflow.processors.LicenseResolver`.
