+++ antenna-basic-assembly/antenna-basic-configuration/pom.xml
@@ -37,18 +37,6 @@
         </dependency>
         <dependency>
             <groupId>org.eclipse.sw360.antenna</groupId>
-            <artifactId>antenna-csv-analyzer</artifactId>
-            <version>${project.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.sw360.antenna</groupId>
-            <artifactId>antenna-json-analyzer</artifactId>
-            <version>${project.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.sw360.antenna</groupId>
             <artifactId>antenna-mvn-dependency-tree-analyzer</artifactId>
             <version>${project.version}</version>
             <scope>runtime</scope>
@@ -74,12 +62,6 @@
         <dependency>
             <groupId>org.eclipse.sw360.antenna</groupId>
             <artifactId>antenna-spdx-license-knowledge-base</artifactId>
-            <version>${project.version}</version>
-            <scope>runtime</scope>
-        </dependency>
-        <dependency>
-            <groupId>org.eclipse.sw360.antenna</groupId>
-            <artifactId>antenna-license-resolver</artifactId>
             <version>${project.version}</version>
             <scope>runtime</scope>
         </dependency>
