+++ antenna/core-workflow-steps/src/main/java/org/eclipse/sw360/antenna/jsonreader/RecordingInputStream.java
