+++ antenna/core-workflow-steps/src/test/resources/JsonReaderTest/data2.json
