+++ antenna/core-workflow-steps/src/test/resources/JsonReaderTest/data3.json
