+++ modules/ort/antenna-ort-result-analyzer/src/main/java/org/eclipse/sw360/antenna/workflow/analyzers/OrtResultAnalyzer.java
