+++ modules/ort/antenna-ort-result-analyzer/src/test/java/org/eclipse/sw360/antenna/workflow/analyzers/OrtResultAnalyzerTest.java
