+++ modules/ort/antenna-ort-result-analyzer/src/test/resources/scan-result-nullAnalyzer.yml
