+++ modules/ort/antenna-ort-result-analyzer/src/test/resources/single_package.yml
